Deparallise build of Text::Markddown::Discount
When relying on just the vendored Discount library in
Text::Markddown::Discount, we also rely on the Makefile of
Text::Markddown::Discount for building the Discount library
which does not seem to have enough dependencies to make it
safe for parallel builds (at least with the default -j12
on my system).  So build the Text::Markddown::Discount
in a single threaded job.

(I think previously the embedded Discount library build was
still failing, but the Homebrew discount package being installed
allowed the link to work anyway.)